+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 ## 1.1.x-dev (2015-05-24)
 
- * Fixed minimum PHP version to >=5.4 (short array syntax)
+**Fixes**
+  * Fixed minimum PHP version to >=5.4 (short array syntax)
+  * Fixed .PDF support
+  
+**Features**
  * Added .ODS to the supported formats
+ * Added 'dataType' property to the 'xlscell' tag. Used to manually override data type of the cell.
+ * Added bundle config 'mewesk_twig_excel.pre_calculate_formulas'. Pre-calculating formulas can be slow in certain cases. Disabling this option can improve the performance but the resulting documents won't show the result of any formulas when opened in a external spreadsheet software.
+ * Added bundle config 'mewesk_twig_excel.disk_caching_directory'. Using disk caching can improve memory consumption by writing data to disk temporarily. Works only for .XLSX and .ODS documents.
+
+**Improvements**
  * Updated to latest stable PhpExcel
- * Updated .travis.yml to cover latest PHP/HHVM and Symfony versions
+ * Updated to latest stable mPDF
  * Improved inline documentation
- * General code clean up
+ * Improved code quality
+ * Added latest PHP/HHVM and Symfony versions to the automated tests
+ * Added more unit tests and functional tests
 
 ## 1.0.0 (2014-03-04)